requirements.txt
discord==2.3.2
neuralintents==0.1.0
python-slugify==8.0.4